move labels to engine package, add skipped field to violations
