+++ demo-output.yaml
@@ -149,6 +149,10 @@
         message: all go files
       - uri: file:///analyzer-lsp/engine/engine_test.go
         message: all go files
+      - uri: file:///analyzer-lsp/engine/labels/labels.go
+        message: all go files
+      - uri: file:///analyzer-lsp/engine/labels/labels_test.go
+        message: all go files
       - uri: file:///analyzer-lsp/examples/golang/main.go
         message: all go files
       - uri: file:///analyzer-lsp/hubapi/violations.go
@@ -168,10 +172,6 @@
       - uri: file:///analyzer-lsp/lsp/protocol/tsprotocol.go
         message: all go files
       - uri: file:///analyzer-lsp/main.go
-        message: all go files
-      - uri: file:///analyzer-lsp/parser/labels/labels.go
-        message: all go files
-      - uri: file:///analyzer-lsp/parser/labels/labels_test.go
         message: all go files
       - uri: file:///analyzer-lsp/parser/rule_parser.go
         message: all go files
